requirements.txt
pyyaml
bleak
requests
pysimplegui
buttplug
pyserial
pybluez
flask
requests-oauthlib
cv2
dxcam
readwritememory
